requirements/production.txt
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.24.2
django-crontab
mysqlclient
python-decouple==3.6
twilio
python-barcode
qrcode
pillow
django-wkhtmltopdf==3.4.0  
openpyxl-image-loader
xhtml2pdf
django-import-export